replaced int with long, now working correctly (63728127 with stopping time of 949) but processing time has more than doubled to ~ 660s

+++ Number.cs
@@ -2,21 +2,21 @@ namespace collatz
 {
 	class Number
 	{
-		private int value;
+		private long value;
 		private int stoppingTime = 0;
 
-		public Number(int startNumber)
+		public Number(long startNumber)
 		{
 			value = startNumber;
 		}
 
-		public Number(int startNumber, int calcStoppingTime) // use to record pre-calcuated numbers
+		public Number(long startNumber, int calcStoppingTime) // use to record pre-calcuated numbers
 		{
 			value = startNumber;
 			stoppingTime = calcStoppingTime;
 		}
 
-		public int getValue()
+		public long getValue()
 		{
 			return this.value;
 		}
